+++ packages/flame/CHANGELOG.md
@@ -12,6 +12,7 @@
  - Add `BaseGame.remove` and `BaseGame.removeAll` helpers for removing components
  - Add `BaseComponent.remove` and `BaseComponent.removeAll` helpers for removing children
  - Rename `Camera.cameraSpeed` to `Camera.speed`
+ - Rename `addShape` to `addHitbox` in `Hitbox` mixin
 
 
 ## [1.0.0-releasecandidate.13]